+++ dali/python/nvidia/dali/_autograph/core/converter_testing.py
@@ -15,7 +15,7 @@
 """Base class for tests in this module."""
 
 import contextlib
-import imp
+import types
 import inspect
 import sys
 
@@ -32,7 +32,7 @@ from nvidia.dali._autograph.utils import hooks
 def allowlist(f):
     """Helper that marks a callable as allowlisted."""
     if "allowlisted_module_for_testing" not in sys.modules:
-        allowlisted_mod = imp.new_module("allowlisted_module_for_testing")
+        allowlisted_mod = types.ModuleType("allowlisted_module_for_testing")
         sys.modules["allowlisted_module_for_testing"] = allowlisted_mod
         config.CONVERSION_RULES = (
             config.DoNotConvert("allowlisted_module_for_testing"),
@@ -76,7 +76,7 @@ class TestingTranspiler(api.PyToLib):
     def get_extra_locals(self):
         retval = super(TestingTranspiler, self).get_extra_locals()
         if self._ag_overrides:
-            modified_ag = imp.new_module("fake_autograph")
+            modified_ag = types.ModuleType("fake_autograph")
             modified_ag.__dict__.update(retval["ag__"].__dict__)
             modified_ag.__dict__.update(self._ag_overrides)
             retval["ag__"] = modified_ag
