+++ dali/test/python/autograph/pyct/test_inspect_utils.py
@@ -17,7 +17,7 @@
 import abc
 import collections
 import functools
-import imp
+import types
 import textwrap
 import unittest
 
@@ -292,8 +292,8 @@ class InspectUtilsTest(unittest.TestCase):
 
     def test_getqualifiedname(self):
         foo = object()
-        qux = imp.new_module("quxmodule")
-        bar = imp.new_module("barmodule")
+        qux = types.ModuleType("quxmodule")
+        bar = types.ModuleType("barmodule")
         baz = object()
         bar.baz = baz
 
@@ -321,7 +321,7 @@ class InspectUtilsTest(unittest.TestCase):
             current_level = []
             for j in range(10):
                 mod_name = "mod_{}_{}".format(i, j)
-                mod = imp.new_module(mod_name)
+                mod = types.ModuleType(mod_name)
                 current_level.append(mod)
                 if i == 9 and j == 9:
                     mod.foo = foo
@@ -347,7 +347,7 @@ class InspectUtilsTest(unittest.TestCase):
         ns = {}
         mods = []
         for i in range(10):
-            mod = imp.new_module("mod_{}".format(i))
+            mod = types.ModuleType("mod_{}".format(i))
             if i == 9:
                 mod.foo = foo
             # Module i refers to module i+1
