+++ dali/test/python/autograph/pyct/test_loader.py
@@ -21,8 +21,10 @@ import textwrap
 import unittest
 
 import gast
+from distutils.version import LooseVersion
 
 from nvidia.dali._autograph.pyct import ast_util
+from nvidia.dali._autograph.pyct import gast_util
 from nvidia.dali._autograph.pyct import loader
 from nvidia.dali._autograph.pyct import parser
 from nvidia.dali._autograph.pyct import pretty_printer
@@ -77,6 +79,7 @@ class LoaderTest(unittest.TestCase):
             decorator_list=[],
             returns=None,
             type_comment=None,
+            **{"type_params": []} if gast_util.get_gast_version() >= LooseVersion("0.5.5") else {},
         )
 
         module, source, _ = loader.load_ast(node)
