+++ dali/test/python/autograph/pyct/test_templates.py
@@ -14,7 +14,7 @@
 # ==============================================================================
 """Tests for templates module."""
 
-import imp
+import types
 import unittest
 
 import gast
@@ -132,7 +132,7 @@ class TemplatesTest(unittest.TestCase):
 
         node = templates.replace(template, foo="b")[0]
         result, _, _ = loader.load_ast(node)
-        mod = imp.new_module("test")
+        mod = types.ModuleType("test")
         mod.b = 3
         self.assertEqual(3, result.test_fn(mod))
