+++ dali/test/python/checkpointing/test_dali_stateless_operators.py
@@ -14,7 +14,6 @@
 
 import os
 import glob
-import nose
 import numpy as np
 import itertools
 import nvidia.dali as dali
@@ -28,7 +27,7 @@ from test_utils import (
     restrict_platform,
 )
 from nose2.tools import params, cartesian_params
-from nose_utils import assert_raises
+from nose_utils import assert_raises, SkipTest
 from nose.plugins.attrib import attr
 
 # Test configuration
@@ -584,7 +583,7 @@ def test_optical_flow_stateless():
     from test_optical_flow import is_of_supported
 
     if not is_of_supported():
-        raise nose.SkipTest("Optical Flow is not supported on this platform")
+        raise SkipTest("Optical Flow is not supported on this platform")
     check_single_sequence_input(fn.optical_flow, "gpu")
