+++ dali/test/python/decoder/test_jpeg_scan_limit.py
@@ -18,9 +18,9 @@ import tempfile
 import unittest
 from nvidia.dali import pipeline_def
 
+from nose_utils import assert_raises
 from nose.plugins.attrib import attr
 from nose2.tools import cartesian_params
-from nose_utils import assert_raises
 
 
 class ProgressiveJpeg(unittest.TestCase):
