+++ dali/test/python/test_external_source_parallel_mxnet.py
@@ -19,8 +19,7 @@
 # to switch between the default numpy and cupy
 
 import mxnet as mx
-from nose import with_setup
-from nose_utils import raises
+from nose_utils import raises, with_setup
 
 from test_pool_utils import setup_function
 from test_external_source_parallel_utils import (
