+++ dali/test/python/test_external_source_parallel_utils.py
@@ -14,8 +14,7 @@
 
 import numpy as np
 import nvidia.dali as dali
-from nose.tools import with_setup
-
+from nose_utils import with_setup
 from test_pool_utils import capture_processes, teardown_function, setup_function
 from test_utils import (
     compare_pipelines,
